more work

+++ astromine-core/src/main/java/com/github/chainmailstudios/astromine/common/recipe/base/EnergyConsumingRecipe.java
@@ -27,5 +27,5 @@ package com.github.chainmailstudios.astromine.common.recipe.base;
 import net.minecraft.inventory.IInventory;
 
 public interface EnergyConsumingRecipe<C extends IInventory> extends AstromineRecipe<C> {
-	double getEnergyConsumed();
+	int getEnergyConsumed();
 }
